pyproject.toml
[project]
name = "headhunter"
version = "0.1.0"
description = ""
authors = [
    {name = "Ruslan Gusencov"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.133.1,<0.134.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)"
]

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=26.1.0,<27.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
